chore(agents): align files with API tag boundaries
Refactor only — no behaviour changes. Both the API client layer and the
providers command package now mirror the OpenAPI spec's tag structure
(one source file per tag), which makes it easier to add a new endpoint
or verb later: a single new file plus its sibling test file.

api/agentstudio/
  client.go      568 LOC monolith → 193 LOC infra-only
                 (Config/Client/NewClient/setHeaders/checkResponse/
                  extractDetail/sentinelFor)
  agents.go      NEW — Agents tag (List/Get/Create/Update/Delete/
                 Publish/Unpublish/Duplicate/InvalidateAgentCache +
                 doAgentMutation/doAgentLifecycle helpers)
  completions.go NEW — Completions tag (Completions + boolToWire)

  agents_test.go (renamed from mutations_test.go) absorbs GetAgent_*
                 from get_agent_test.go (deleted) and the ListAgents
                 happy-path tests previously in client_test.go.
  client_test.go now holds only the cross-cutting infra tests
                 (NewClient validation, checkResponse error mapping,
                 ctx cancellation, header omission). The error-mapping
                 suite still uses ListAgents as a vehicle since it's
                 the simplest GET shape; the file header explains why.

pkg/cmd/agents/providers/
  providers.go   738 LOC → 93 LOC (parent + readBody/ctxOrBackground/
                 relTimeOrDash helpers)
  list.go, get.go, create.go, update.go, delete.go, models.go
                 NEW — one file per verb, same package so MaskInput
                 and the helpers stay un-exported

  providers_test.go shrinks to the test harness (newClientForServer +
                 writeTempJSON); each verb gets a sibling _test.go
                 with its own scenarios.

AGENTS.md gains a "File organisation" section documenting both layers'
tag-aligned conventions and the explicit decision NOT to promote
sub-group verbs to their own subpackages.

All gates green: gofumpt, go vet, go test ./..., e2e TestAgents/{cache,
try,dry-run,config,providers}, task build, golangci-lint on touched
dirs.
