+++ api/agentstudio/agents_test.go
@@ -1,5 +1,11 @@
 package agentstudio
 
+// Tests for the Agents API tag — the methods defined in agents.go.
+//
+// Cross-cutting concerns (NewClient validation, error mapping via
+// checkResponse, header injection) live in client_test.go and use
+// ListAgents as a vehicle since it's the simplest endpoint shape.
+
 import (
 	"context"
 	"encoding/json"
@@ -11,6 +17,130 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+func TestListAgents_Success(t *testing.T) {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/1/agents", func(w http.ResponseWriter, r *http.Request) {
+		assert.Equal(t, http.MethodGet, r.Method)
+		assert.Equal(t, "APP123", r.Header.Get(HeaderApplicationID))
+		assert.Equal(t, "key-abc", r.Header.Get(HeaderAPIKey))
+		assert.Equal(t, "cli-test", r.Header.Get(HeaderUserID))
+		assert.Equal(t, "application/json", r.Header.Get("Accept"))
+
+		assert.Equal(t, "2", r.URL.Query().Get("page"))
+		assert.Equal(t, "25", r.URL.Query().Get("limit"))
+		assert.Equal(t, "prov-1", r.URL.Query().Get("providerId"))
+
+		require.NoError(t, json.NewEncoder(w).Encode(PaginatedAgentsResponse{
+			Data: []Agent{{
+				ID:           "11111111-1111-1111-1111-111111111111",
+				Name:         "Concierge",
+				Status:       StatusDraft,
+				Instructions: "Be helpful.",
+			}},
+			Pagination: PaginationMetadata{
+				Page: 2, Limit: 25, TotalCount: 1, TotalPages: 1,
+			},
+		}))
+	})
+
+	_, c := newTestClient(t, mux)
+
+	got, err := c.ListAgents(context.Background(), ListAgentsParams{
+		Page:       2,
+		Limit:      25,
+		ProviderID: "prov-1",
+	})
+	require.NoError(t, err)
+	require.Len(t, got.Data, 1)
+	assert.Equal(t, "Concierge", got.Data[0].Name)
+	assert.Equal(t, StatusDraft, got.Data[0].Status)
+	assert.Equal(t, 25, got.Pagination.Limit)
+}
+
+func TestListAgents_OmitsZeroParams(t *testing.T) {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/1/agents", func(w http.ResponseWriter, r *http.Request) {
+		assert.Empty(t, r.URL.RawQuery, "expected no query params for zero-valued params")
+		_, _ = w.Write([]byte(`{"data":[],"pagination":{"page":1,"limit":10,"totalCount":0,"totalPages":0}}`))
+	})
+
+	_, c := newTestClient(t, mux)
+	_, err := c.ListAgents(context.Background(), ListAgentsParams{})
+	require.NoError(t, err)
+}
+
+func TestGetAgent_Success(t *testing.T) {
+	mux := http.NewServeMux()
+	mux.HandleFunc(
+		"/1/agents/abc-123",
+		func(w http.ResponseWriter, r *http.Request) {
+			assert.Equal(t, http.MethodGet, r.Method)
+			assert.Equal(t, "APP123", r.Header.Get(HeaderApplicationID))
+			_, _ = w.Write([]byte(`{
+				"id":"abc-123",
+				"name":"Concierge",
+				"status":"published",
+				"instructions":"Be helpful.",
+				"createdAt":"2025-01-02T03:04:05Z"
+			}`))
+		},
+	)
+
+	_, c := newTestClient(t, mux)
+
+	got, err := c.GetAgent(context.Background(), "abc-123")
+	require.NoError(t, err)
+	assert.Equal(t, "abc-123", got.ID)
+	assert.Equal(t, "Concierge", got.Name)
+	assert.Equal(t, StatusPublished, got.Status)
+}
+
+func TestGetAgent_EmptyIDRejected(t *testing.T) {
+	_, c := newTestClient(t, http.NewServeMux())
+
+	_, err := c.GetAgent(context.Background(), "  ")
+	require.Error(t, err)
+	assert.Contains(t, err.Error(), "agent id is required")
+}
+
+func TestGetAgent_NotFound(t *testing.T) {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/1/agents/missing", func(w http.ResponseWriter, _ *http.Request) {
+		w.WriteHeader(http.StatusNotFound)
+		_, _ = w.Write([]byte(`{"detail":"Agent not found"}`))
+	})
+
+	_, c := newTestClient(t, mux)
+
+	_, err := c.GetAgent(context.Background(), "missing")
+	require.Error(t, err)
+	assert.True(t, errors.Is(err, ErrNotFound), "expected ErrNotFound, got %v", err)
+}
+
+func TestGetAgent_PathEscapesID(t *testing.T) {
+	// Confirms IDs with reserved chars are encoded into the path so they
+	// can't poison the URL (e.g., "../" attempts).
+	mux := http.NewServeMux()
+	mux.HandleFunc(
+		"/1/agents/weird%2Fid",
+		func(w http.ResponseWriter, _ *http.Request) {
+			_, _ = w.Write([]byte(`{
+				"id":"weird/id",
+				"name":"x",
+				"status":"draft",
+				"instructions":"",
+				"createdAt":"2025-01-01T00:00:00Z"
+			}`))
+		},
+	)
+
+	_, c := newTestClient(t, mux)
+
+	got, err := c.GetAgent(context.Background(), "weird/id")
+	require.NoError(t, err)
+	assert.Equal(t, "weird/id", got.ID)
+}
 
 func TestCreateAgent_Success(t *testing.T) {
 	wantBody := `{"name":"Concierge","instructions":"Be helpful."}`
