+++ pkg/cmd/agents/providers/get_test.go
@@ -1,0 +1,41 @@
+package providers
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/algolia/cli/test"
+)
+
+func Test_runGetCmd_MasksByDefault(t *testing.T) {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/1/providers/p1", func(w http.ResponseWriter, _ *http.Request) {
+		_, _ = w.Write([]byte(`{
+			"id":"p1","name":"openai-prod","providerName":"openai",
+			"input":{"apiKey":"sk-LEAKED"},
+			"createdAt":"2026-01-01T00:00:00Z","updatedAt":"2026-01-01T00:00:00Z"
+		}`))
+	})
+	ts := httptest.NewServer(mux)
+	t.Cleanup(ts.Close)
+
+	f, out := test.NewFactory(false, nil, nil, "")
+	f.AgentStudioClient = newClientForServer(t, ts)
+
+	cmd := NewProvidersCmd(f)
+	result, err := test.Execute(cmd, "get p1", out)
+	require.NoError(t, err)
+	assert.NotContains(t, result.String(), "sk-LEAKED")
+	assert.Contains(t, result.String(), `"apiKey":"***"`)
+}
+
+func Test_runGetCmd_RequiresProviderID(t *testing.T) {
+	f, out := test.NewFactory(false, nil, nil, "")
+	cmd := NewProvidersCmd(f)
+	_, err := test.Execute(cmd, "get", out)
+	require.Error(t, err)
+}
